factor out the command button into an interaction component
